Метод получения дамага в Player Controller

+++ WGA Project/Assets/GameSettings.cs
@@ -17,5 +17,5 @@ public static class GameSettings
     public static int healtRaft=0;
     public static int healthFrog=0;
     public static int playerScore;
-    public static int playerHP;
+    public static float playerHP;
 }
